Splitting out fs3 functions.
